Major refactoring to support the needed functionality

+++ src/main/java/com/opdapp/dto/DrugPackageDTO.java
@@ -2,7 +2,7 @@ package com.opdapp.dto;
 
 public class DrugPackageDTO {
 
-    private int drugPackageId;
+    private long drugPackageId;
 
     private DrugDTO drug;
 
@@ -14,11 +14,11 @@ public class DrugPackageDTO {
 
     private double unitPrice;
 
-    public int getDrugPackageId() {
+    public long getDrugPackageId() {
         return drugPackageId;
     }
 
-    public void setDrugPackageId(int drugPackageId) {
+    public void setDrugPackageId(long drugPackageId) {
         this.drugPackageId = drugPackageId;
     }
